@@ -4,62 +4,50 @@
     "scripts": {
         "ng": "ng",
         "start": "ng serve",
+        "start:dev": "ng serve -c dev",
         "build": "ng build",
         "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
+        "lint": "ng lint"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "~12.2.3",
-        "@angular/cdk": "^11.2.11",
-        "@angular/common": "~12.2.3",
-        "@angular/compiler": "~12.2.3",
-        "@angular/core": "~12.2.3",
-        "@angular/forms": "~12.2.3",
-        "@angular/platform-browser": "~12.2.3",
-        "@angular/platform-browser-dynamic": "~12.2.3",
-        "@angular/router": "~12.2.3",
-        "@fullcalendar/core": "4.1.0",
-        "@fullcalendar/daygrid": "4.1.0",
-        "@fullcalendar/interaction": "4.1.0",
-        "@fullcalendar/timegrid": "4.1.0",
-        "@ng-bootstrap/ng-bootstrap": "9.1.0",
-        "bootstrap": "4.3.1",
-        "chart.js": "2.7.3",
-        "core-js": "^2.6.9",
-        "jquery": ">=3.5.0",
-        "popper": "^1.0.1",
-        "primeflex": "2.0.0",
-        "primeicons": "^4.1.0",
-        "primeng": "^11.4.5",
-        "prismjs": "1.20.0",
-        "rxjs": "6.6.7",
-        "rxjs-compat": "6.3.3",
+        "@angular/animations": "15.2.0",
+        "@angular/cdk": "15.2.9",
+        "@angular/common": "15.2.0",
+        "@angular/compiler": "15.2.0",
+        "@angular/core": "15.2.0",
+        "@angular/forms": "15.2.0",
+        "@angular/platform-browser": "15.2.0",
+        "@angular/platform-browser-dynamic": "15.2.0",
+        "@angular/router": "15.2.0",
+        "@fullcalendar/angular": "^6.1.9",
+        "@fullcalendar/core": "^6.1.9",
+        "@fullcalendar/daygrid": "6.1.9",
+        "@fullcalendar/interaction": "6.1.9",
+        "@fullcalendar/timegrid": "6.1.9",
+        "chart.js": "^3.3.2",
+        "primeflex": "^3.2.1",
+        "primeicons": "6.0.1",
+        "primeng": "15.2.0",
+        "prismjs": "1.9.0",
+        "quill": "^1.3.7",
+        "rxjs": "~7.8.0",
         "sass": "1.27.0",
         "sass-loader": "10.0.5",
-        "tslib": "^2.0.0",
-        "zone.js": "~0.11.4"
+        "tslib": "^2.3.0",
+        "zone.js": "~0.12.0"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~12.2.3",
-        "@angular/cli": "~12.2.3",
-        "@angular/compiler-cli": "~12.2.3",
-        "@angular/language-service": "~12.2.3",
-        "@types/jasmine": "~3.8.2",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "^12.11.1",
-        "codelyzer": "^6.0.0",
-        "jasmine-core": "~3.9.0",
-        "jasmine-spec-reporter": "~5.0.0",
-        "karma": "~6.3.2",
+        "@angular-devkit/build-angular": "15.2.0",
+        "@angular/cli": "15.2.0",
+        "@angular/compiler-cli": "15.2.0",
+        "@types/jasmine": "~4.3.0",
+        "jasmine-core": "~4.5.0",
+        "karma": "~6.4.0",
         "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage-istanbul-reporter": "~3.0.2",
-        "karma-jasmine": "~4.0.0",
-        "karma-jasmine-html-reporter": "^1.5.0",
-        "protractor": "~7.0.0",
-        "ts-node": "~7.0.0",
-        "tslint": "~6.1.0",
-        "typescript": "~4.3.5"
+        "karma-coverage": "~2.2.0",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.0.0",
+        "typescript": "~4.9.4"
     }
 }
